Gameplay: Refactor out all api calls to a separate module
- pubnub calls and calls to rails server
